requirements.txt
flask==3.0.3
anthropic==0.34.2
pdfplumber==0.11.4
reportlab==4.2.2
pillow==10.4.0
numpy==1.26.4
gunicorn==22.0.0
